Update c2py converters
- move bravais_lattice::point_t converter into its own header
- avoid including any umbrella headers like triqs/gfs.hpp which causes problems
